+++ NEWS.md
@@ -9,6 +9,11 @@
     equal up to a rotation.
     * Functions `get_criterion` and `plot_criterion` to get and plot the criterion used
     for model selection easilly.
+* New feature:
+    * Function PhyloEM can now allow for negative values of alpha, to fit an EB model.
+* Bug fix:
+    * In function plot, fixed bug that led to the wrong parameters being selected when 
+    doing ancestral trait reconstruction.
 
 # PhylogeneticEM 1.1.0
 * Deprecation Fix: